+++ test/finch_debugging_test.dart
@@ -1,9 +1,9 @@
+import 'package:finch/src/tools/http/http.dart';
 import 'package:test/test.dart';
 import 'package:finch/src/tools/console.dart';
 import 'package:finch/finch_route.dart';
 import 'package:finch/finch_app.dart';
 import 'package:finch/finch_tools.dart';
-import 'package:http/http.dart' as http;
 
 void main() async {
   FinchApp server = FinchApp(
@@ -100,21 +100,21 @@ void main() async {
 
   group("test debugger", () {
     test("checking Debugger1", () async {
-      var req = await http.get(
+      var req = await FinchHttp.get(
         Uri.parse("http://localhost:${httpServer.port}/debugger/console.js"),
       );
 
       expect(
-        req.statusCode,
+        req.status,
         200,
         reason:
             "Response success should contain 'console.js' it used for debugging",
       );
-      expect(req.statusCode, 200, reason: "Status code should be 200");
+      expect(req.status, 200, reason: "Status code should be 200");
     });
 
     test("checking Debugger2", () async {
-      var req = await http.get(
+      var req = await FinchHttp.get(
         Uri.parse("http://localhost:${httpServer.port}/debug_test/"),
       );
       var content = req.body;
@@ -125,11 +125,11 @@ void main() async {
         reason:
             "Response success should contain '<script src=...>' used for debugging",
       );
-      expect(req.statusCode, 200, reason: "Status code should be 200");
+      expect(req.status, 200, reason: "Status code should be 200");
     });
 
     test("checking Debugger3", () async {
-      var req = await http.get(
+      var req = await FinchHttp.get(
         Uri.parse("http://localhost:${httpServer.port}/api/info"),
       );
       var content = req.body;
@@ -139,7 +139,7 @@ void main() async {
         false,
         reason: "debugger used only on HTML pages",
       );
-      expect(req.statusCode, 200, reason: "Status code should be 200");
+      expect(req.status, 200, reason: "Status code should be 200");
     });
   });
 }